Made so level management actually works

+++ SoundShapesServer/Endpoints/Api/Moderation/ApiManageUserEndpoints.cs
@@ -1,5 +1,6 @@
 using System.Net;
 using System.Reflection.Metadata.Ecma335;
+using Bunkum.CustomHttpListener.Parsing;
 using Bunkum.HttpServer;
 using Bunkum.HttpServer.Endpoints;
 using Bunkum.HttpServer.Responses;
@@ -12,19 +13,21 @@ namespace SoundShapesServer.Endpoints.Api.Moderation;
 
 public class ApiManageUserEndpoints : EndpointGroup
 {
-    [ApiEndpoint("user/{id}/remove")]
+    [ApiEndpoint("user/{id}/remove", Method.Post)]
     public Response RemoveUser(RequestContext context, RealmDatabaseContext database, GameUser user, string id)
     {
         if (PermissionHelper.IsUserAdmin(user) == false) return HttpStatusCode.Forbidden;
 
         GameUser? userToRemove = database.GetUserWithId(id);
         if (userToRemove == null) return HttpStatusCode.NotFound;
+
+        if (userToRemove.PermissionsType >= user.PermissionsType) return HttpStatusCode.Unauthorized;
         
         database.RemoveUser(userToRemove);
         return HttpStatusCode.OK;
     }
 
-    [ApiEndpoint("user/{id}/setPermissions")]
+    [ApiEndpoint("user/{id}/setPermissions", Method.Post)]
     public Response SetUserPermissions(RequestContext context, RealmDatabaseContext database, GameUser user, string id, ApiSetUserPermissionsRequest body)
     {
         if (PermissionHelper.IsUserAdmin(user) == false) return HttpStatusCode.Forbidden;
@@ -36,7 +39,7 @@ public class ApiManageUserEndpoints : EndpointGroup
         if (body.PermissionsType > user.PermissionsType || userToSetPermissionsOf.PermissionsType >= user.PermissionsType) return HttpStatusCode.Unauthorized;
         
         
-        database.SetUserPermissions(user, type);
+        database.SetUserPermissions(userToSetPermissionsOf, type);
         return HttpStatusCode.OK;
     }
 }
